+++ __tests__/v2/CreateDatasetModal.test.js
@@ -1,5 +1,6 @@
 import { describe, it, beforeEach, expect, vi } from 'vitest';
-import { fireEvent, render } from '@testing-library/svelte';
+import { render } from '@testing-library/svelte';
+import userEvent from '@testing-library/user-event';
 
 // Mocking fetch
 global.fetch = vi.fn();
@@ -13,7 +14,7 @@ vi.mock('$app/state', () => {
 			},
 			data: {
 				userInfo: {
-					project_dirs: ['/path/to/proj/dir']
+					project_dirs: ['/path/to/proj/dir1', '/path/to/proj/dir2']
 				}
 			}
 		}
@@ -37,7 +38,13 @@ global.window.bootstrap = {
 import CreateDatasetModal from '../../src/lib/components/v2/projects/datasets/CreateDatasetModal.svelte';
 
 const defaultProps = {
-	props: { createDatasetCallback: vi.fn() }
+	props: {
+		project: /** @type {import("fractal-components/types/api").ProjectV2} */ ({
+			id: 1,
+			name: 'My Project'
+		}),
+		createDatasetCallback: vi.fn()
+	}
 };
 
 function mockFetch() {
@@ -59,98 +66,135 @@ describe('CreateDatasetModal', () => {
 	});
 
 	it('validate missing name', async () => {
+		const user = userEvent.setup();
 		mockFetch();
 		const result = render(CreateDatasetModal, defaultProps);
-		await fireEvent.click(result.getByRole('button', { name: 'Save' }));
+		await user.click(result.getByRole('button', { name: 'Save' }));
 		expect(result.queryAllByText('Required field').length).eq(1);
 	});
 
-	it('create dataset with string filter', async () => {
+	it('validate absolute zarr subfolder', async () => {
+		const user = userEvent.setup();
 		mockFetch();
-		const createDatasetCallback = vi.fn();
-		const result = render(CreateDatasetModal, {
-			props: { createDatasetCallback }
-		});
-		await fireEvent.input(result.getByRole('textbox', { name: 'Dataset Name' }), {
-			target: { value: 'my dataset' }
-		});
-		await fireEvent.input(result.getByRole('textbox', { name: 'Zarr dir' }), {
-			target: { value: '/tmp' }
-		});
-		await fireEvent.click(result.getByRole('button', { name: 'Save' }));
+		const result = render(CreateDatasetModal, defaultProps);
+		await user.click(result.getByRole('button', { name: 'Save' }));
+		await user.type(result.getByRole('textbox', { name: 'Zarr subfolder' }), '/my_subfolder');
+		expect(result.queryAllByText('Zarr subfolder must be a relative path').length).eq(1);
+	});
+
+	it('create dataset specifying Zarr subfolder', async () => {
+		const user = userEvent.setup();
+		mockFetch();
+		const result = render(CreateDatasetModal, defaultProps);
+		await user.type(result.getByRole('textbox', { name: 'Dataset Name' }), 'my dataset');
+		await user.type(result.getByRole('textbox', { name: 'Zarr subfolder' }), 'my_subfolder');
+		expect(await result.findByText('/path/to/proj/dir1/my_subfolder')).toBeVisible();
+		await user.click(result.getByRole('button', { name: 'Save' }));
 		expect(fetch).toHaveBeenLastCalledWith(
 			'/api/v2/project/1/dataset',
 			expect.objectContaining({
 				body: JSON.stringify({
 					name: 'my dataset',
-					zarr_dir: '/tmp'
+					project_dir: '/path/to/proj/dir1',
+					zarr_subfolder: 'my_subfolder'
 				})
 			})
 		);
 	});
 
-	it('create dataset without specifying zarr dir', async () => {
+	it('create dataset selecting the second project dir', async () => {
+		const user = userEvent.setup();
 		mockFetch();
-		const createDatasetCallback = vi.fn();
-		const result = render(CreateDatasetModal, {
-			props: { createDatasetCallback }
-		});
-		await fireEvent.input(result.getByRole('textbox', { name: 'Dataset Name' }), {
-			target: { value: 'my dataset' }
-		});
-		await fireEvent.click(result.getByRole('button', { name: 'Save' }));
+		const result = render(CreateDatasetModal, defaultProps);
+		await user.type(result.getByRole('textbox', { name: 'Dataset Name' }), 'my dataset');
+		await user.selectOptions(result.getByRole('combobox', { name: 'Project directory' }), [
+			'/path/to/proj/dir2'
+		]);
+		await user.type(result.getByRole('textbox', { name: 'Zarr subfolder' }), 'my_subfolder');
+		expect(await result.findByText('/path/to/proj/dir2/my_subfolder')).toBeVisible();
+		await user.click(result.getByRole('button', { name: 'Save' }));
 		expect(fetch).toHaveBeenLastCalledWith(
 			'/api/v2/project/1/dataset',
 			expect.objectContaining({
 				body: JSON.stringify({
-					name: 'my dataset'
+					name: 'my dataset',
+					project_dir: '/path/to/proj/dir2',
+					zarr_subfolder: 'my_subfolder'
+				})
+			})
+		);
+	});
+
+	it('create dataset without specifying zarr subfolder', async () => {
+		const user = userEvent.setup();
+		mockFetch();
+		const result = render(CreateDatasetModal, defaultProps);
+		await user.type(result.getByRole('textbox', { name: 'Dataset Name' }), 'my dataset');
+		expect(
+			await result.findByText('/path/to/proj/dir1/fractal/1_my_project/{dataset_id}_my_dataset')
+		).toBeVisible();
+		await user.click(result.getByRole('button', { name: 'Save' }));
+		expect(fetch).toHaveBeenLastCalledWith(
+			'/api/v2/project/1/dataset',
+			expect.objectContaining({
+				body: JSON.stringify({
+					name: 'my dataset',
+					project_dir: '/path/to/proj/dir1'
 				})
 			})
 		);
 	});
 
 	it('validate missing dataset name', async () => {
+		const user = userEvent.setup();
 		const result = render(CreateDatasetModal, defaultProps);
-		expect(await result.findByText('/path/to/proj/dir')).toBeVisible();
-		await fireEvent.click(result.getByRole('button', { name: 'Save' }));
+		await user.click(result.getByRole('button', { name: 'Save' }));
 		expect(result.queryAllByText('Required field').length).eq(1);
 	});
 
-	it('display invalid zarr dir error', async () => {
+	it('display validation errors from server', async () => {
+		const user = userEvent.setup();
+
 		/** @type {import('vitest').Mock} */ (fetch).mockResolvedValueOnce({
 			ok: false,
 			status: 422,
 			json: async () => ({
 				detail: [
 					{
-						loc: ['body', 'zarr_dir'],
-						msg: "URLs must begin with '/' or 's3'.",
+						loc: ['body', 'name'],
+						msg: 'validation error 1',
+						type: 'value_error'
+					},
+					{
+						loc: ['body', 'project_dir'],
+						msg: 'validation error 2',
+						type: 'value_error'
+					},
+					{
+						loc: ['body', 'zarr_subfolder'],
+						msg: 'validation error 3',
 						type: 'value_error'
 					}
 				]
 			})
 		});
 
-		const createDatasetCallback = vi.fn();
-		const result = render(CreateDatasetModal, {
-			props: { createDatasetCallback }
-		});
-		await fireEvent.input(result.getByRole('textbox', { name: 'Dataset Name' }), {
-			target: { value: 'my dataset' }
-		});
-		await fireEvent.input(result.getByRole('textbox', { name: 'Zarr dir' }), {
-			target: { value: 'foo' }
-		});
-		await fireEvent.click(result.getByRole('button', { name: 'Save' }));
+		const result = render(CreateDatasetModal, defaultProps);
+		await user.type(result.getByRole('textbox', { name: 'Dataset Name' }), 'my dataset');
+		await user.type(result.getByRole('textbox', { name: 'Zarr subfolder' }), 'foo');
+		await user.click(result.getByRole('button', { name: 'Save' }));
 		expect(fetch).toHaveBeenLastCalledWith(
 			'/api/v2/project/1/dataset',
 			expect.objectContaining({
 				body: JSON.stringify({
 					name: 'my dataset',
-					zarr_dir: 'foo'
+					project_dir: '/path/to/proj/dir1',
+					zarr_subfolder: 'foo'
 				})
 			})
 		);
-		expect(await result.findByText(/URLs must begin with '\/' or 's3'/)).toBeVisible();
+		expect(await result.findByText(/validation error 1/)).toBeVisible();
+		expect(await result.findByText(/validation error 2/)).toBeVisible();
+		expect(await result.findByText(/validation error 3/)).toBeVisible();
 	});
 });