+++ src/lib/components/v2/workflow/RunWorkflowModal.svelte
@@ -15,6 +15,8 @@
 		STATUS_KEY
 	} from '$lib/common/workflow_utilities';
 	import { normalizePayload } from 'fractal-components';
+	import { splitZarrDir } from '$lib/common/component_utilities';
+	import { page } from '$app/state';
 
 	/**
 	 * @typedef {Object} Props
@@ -173,6 +175,14 @@
 		const { id, name, zarr_dir } = /** @type {import('fractal-components/types/api').DatasetV2} */ (
 			selectedDataset
 		);
+
+		const { projectDir } = splitZarrDir(zarr_dir, page.data.userInfo.project_dirs);
+		if (projectDir === '') {
+			throw new Error(
+				`You cannot reset dataset "${selectedDataset?.name}", because its Zarr directory is not part of your project directories. Create a fresh dataset instead to rerun the workflow on.`
+			);
+		}
+
 		await handleDatasetDelete(id);
 		const newDatasets = datasets.filter((d) => d.id !== id);
 		const newDataset = await handleDatasetCreate(name, zarr_dir);
@@ -196,13 +206,15 @@
 	async function handleDatasetCreate(datasetName, zarrDir) {
 		const headers = new Headers();
 		headers.set('Content-Type', 'application/json');
+		const { projectDir, zarrSubfolder } = splitZarrDir(zarrDir, page.data.userInfo.project_dirs);
 		const response = await fetch(`/api/v2/project/${workflow.project_id}/dataset`, {
 			method: 'POST',
 			credentials: 'include',
 			headers,
 			body: normalizePayload({
 				name: datasetName,
-				zarr_dir: zarrDir
+				project_dir: projectDir,
+				zarr_subfolder: zarrSubfolder
 			})
 		});
 		if (!response.ok) {