+++ tests/v2/images.spec.js
@@ -5,7 +5,8 @@ test('Dataset images [v2]', async ({ page, project }) => {
 	await page.waitForURL(project.url);
 	await waitPageLoading(page);
 
-	const randomPath = `/tmp/${Math.random().toString(36).substring(7)}`;
+	const randomZarrSubfolder = Math.random().toString(36).substring(7);
+	const randomPath = `/tmp/${randomZarrSubfolder}`;
 	const modal = page.locator('.modal.show');
 
 	await test.step('Create test dataset', async () => {
@@ -14,7 +15,8 @@ test('Dataset images [v2]', async ({ page, project }) => {
 		await modal.waitFor();
 		await modal.getByRole('textbox', { name: 'Dataset Name' }).fill('test-dataset');
 		await modal.getByRole('button', { name: 'Advanced options' }).click();
-		await modal.getByRole('textbox', { name: 'Zarr dir' }).fill(randomPath);
+		await modal.getByRole('combobox', { name: 'Project directory' }).selectOption('/tmp');
+		await modal.getByRole('textbox', { name: 'Zarr subfolder' }).fill(randomZarrSubfolder);
 		await modal.getByRole('button', { name: 'Save' }).click();
 		await waitModalClosed(page);
 	});