+++ tests/v2/user_utils.js
@@ -3,16 +3,17 @@ import { selectSlimSelect, waitModalClosed, waitPageLoading } from '../utils';
 
 /**
  * @param {import('@playwright/test').Page} page
+ * @param {string} projectDir
  * @returns {Promise<string>}
  */
-export async function createTestUser(page) {
+export async function createTestUser(page, projectDir = '/tmp') {
 	const randomEmail = Math.random().toString(36).substring(7) + '@example.com';
 	await page.goto('/v2/admin/users/register');
 	await waitPageLoading(page);
 	await page.getByRole('textbox', { name: 'E-mail' }).fill(randomEmail);
 	await page.getByLabel('Password', { exact: true }).fill('test');
 	await page.getByLabel('Confirm password').fill('test');
-	await page.getByRole('textbox', { name: 'Project dir' }).fill('/tmp');
+	await page.getByRole('textbox', { name: 'Project dir' }).fill(projectDir);
 	await page.getByRole('button', { name: 'Save' }).click();
 	await page.waitForURL(/\/v2\/admin\/users\/\d+\/edit/);
 	await waitPageLoading(page);